Rename blockprocessor file to block_processor for consistency
