+++ nano/core_test/block_processor.cpp
