+++ nano/node/local_block_broadcaster.hpp
@@ -5,7 +5,7 @@
 #include <nano/lib/locks.hpp>
 #include <nano/lib/processing_queue.hpp>
 #include <nano/lib/rate_limiting.hpp>
-#include <nano/node/blockprocessor.hpp>
+#include <nano/node/block_processor.hpp>
 #include <nano/node/fwd.hpp>
 #include <nano/secure/common.hpp>
